動画再生速度を0.5fps（1分）に変更 / float fps対応
- main.py: --fps 10 → 0.5（30フレーム = 60秒）
- generate_video.py: fps type int → float（0.5fps対応）

+++ visualization/generate_video.py
@@ -304,7 +304,7 @@ def main():
     parser.add_argument("data_dir", type=str, help="データディレクトリのパス")
     parser.add_argument("-o", "--output", type=str, default=None,
                         help="出力MP4ファイル名 (デフォルト: <data_dir>/simulation.mp4)")
-    parser.add_argument("--fps", type=int, default=10,
+    parser.add_argument("--fps", type=float, default=10,
                         help="フレームレート (デフォルト: 10)")
     parser.add_argument("--dpi", type=int, default=150,
                         help="DPI (デフォルト: 150)")
